requirements.txt
pytest
requests
pytest-html
pytest-xdist
python-dotenv
pytest-rerunfailures
allure-pytest